refactor: extract monitor polling and status
