+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Format follows [Keep a Changelog](https://keepachangelog.com/). This project use
 ### Changed
 - Split local wire decoding, BLE transport, and cloud auth-key retrieval out of `local_transport.py` into responsibility-focused `protocol.py`, `ble_transport.py`, and `cloud_api.py` modules without changing transport behavior (#66).
 - Split monitor alert delivery, rules automation, shutdown-restore policy, and shared telemetry helpers into focused modules while keeping `PecronMonitor` behavior and method contracts unchanged (#66).
+- Split local transport setup/status polling and one-shot status rendering into `monitor_polling.py` and `monitor_status.py`, preserving retry timing and CLI output contracts (#66).
 
 ### Fixed
 - Continuous local monitoring now retries incomplete E3600/E3800 multi-packet reads within the current poll cycle, targets only devices still missing telemetry, and advances to the next future cycle boundary when retries overrun an interval (#88).
