+++ pyproject.toml
@@ -40,8 +40,10 @@ py-modules = [
     "monitor",
     "monitor_alerts",
     "monitor_data",
+    "monitor_polling",
     "monitor_restore",
     "monitor_rules",
+    "monitor_status",
     "pecron_monitor",
     "protocol",
     "setup_wizard",
@@ -80,8 +82,10 @@ source = [
     "monitor",
     "monitor_alerts",
     "monitor_data",
+    "monitor_polling",
     "monitor_restore",
     "monitor_rules",
+    "monitor_status",
     "pecron_monitor",
     "protocol",
     "setup_wizard",
